+++ src/test/java/roomescape/acceptance/AdminViewAcceptanceTest.java
@@ -1,0 +1,24 @@
+package roomescape.acceptance;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+import org.springframework.http.HttpStatus;
+
+class AdminViewAcceptanceTest extends BaseAcceptanceTest {
+
+    @ParameterizedTest
+    @ValueSource(strings = {"/admin", "/admin/reservation", "/admin/time"})
+    void adminPage(String path) {
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .when().get(path)
+                .then().log().all()
+                .extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+}
